Tests/LibWeb/Text/expected/DOM/Elements-from-point-3.txt

 Some text  Elements at point 30, 20: p < div < body < html   
